src/requirements.txt
pydantic-settings==2.6.1
pandas==2.2.3
pydantic==2.11.3
fastapi==0.115.12
uvicorn==0.34.0
pyannote.audio==3.3.2
torch==2.6.0
pydub==0.25.1
openai-whisper==20240930
yt_dlp==2024.12.6
crewai==0.114.0
edge_tts==7.0.1
path==17.1.0
python-multipart==0.0.19
nltk==3.9.1
wordcloud==1.9.4
matplotlib==3.9.2